+++ src/games/candy-crush/src/objects/tile.ts
@@ -1,0 +1,13 @@
+import { IImageConstructor } from '../interfaces/image.interface';
+
+export class Tile extends Phaser.GameObjects.Image {
+  constructor(aParams: IImageConstructor) {
+    super(aParams.scene, aParams.x, aParams.y, aParams.texture, aParams.frame);
+
+    // set image settings
+    this.setOrigin(0, 0);
+    this.setInteractive();
+
+    this.scene.add.existing(this);
+  }
+}