+++ CHANGELOG.md
@@ -28,6 +28,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Refactor `alpha-adjust`
 - Refactor `asteroid`
 - Refactor `blockade`
+- Refactor `candy-crush`
 
 ---